controls from CenterBowl moved to its own component, placed above CenterBowl in Configurator.tsx
